Renamed mesh src and includes
